Updated all screens to typescript : initial with any
