+++ front-end/tsconfig.json
@@ -1,6 +1,7 @@
 {
    "compilerOptions": {
       "target": "es5",
+      "downlevelIteration": true,
       "lib": ["dom", "dom.iterable", "esnext"],
       "allowJs": true,
       "skipLibCheck": true,
